backend/pyproject.toml
[tool.poetry]
name = "journiq"
version = "0.1.0"
description = ""
authors = ["wayneuconn <wayne.uconn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
google-cloud-firestore = "^2.16.0"
websockets = "^12.0"
asyncpg = "^0.29.0"
sqlalchemy = "^2.0.30"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
